requirements.txt
discord.py==2.3.2
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1
anthropic==0.39.0
